update inner class and top level class
